+++ R/spectrum_correlation.R
@@ -58,8 +58,8 @@ spectrum_correlation <- function(X, parallel = 1, pb = TRUE, method = 1, cor.met
   
   
   # is extended sel tab
-  if (!warbleR::is_extended_selection_table(X)) 
-    stop2("'X' must be and extended selection table")
+  # if (!warbleR::is_extended_selection_table(X)) 
+    # stop2("'X' must be and extended selection table")
   
   # If parallel is not numeric
   if (!is.numeric(parallel)) stop2("'parallel' must be a numeric vector of length 1")